requirements.txt
asgiref==3.5.0
Django==4.0.1
psycopg2-binary==2.9.3
python-slugify==5.0.2
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
transliterate==1.10.2
